pyproject.toml
[tool.poetry]
name = "PatchGuru"
version = "0.0.1"
description = "PatchGuru is an LLM-powered tool that automatically infer patch oracles from natural language artifacts in pull requests and utilizes them to detect the inconsistencies between code changes and their corresponding descriptions."
authors = ["Thanh Le-Cong <thanhcls1316@gmail.com>"]
repository = "https://github.com/thanhlecongg/PatchGuru"
documentation = "https://thanhlecongg.github.io/PatchGuru/"
readme = "README.md"
packages = [
  {include = "patchguru"}
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
colorlog = "^6.9.0"
GitPython = "^3.1.45"
pydantic = "^2.11.9"
unidiff = "^0.7.5"
libcst = "^1.8.4"
PyGithub = "^2.8.1"
openai = "^1.108.1"
docker = "^7.1.0"
numpy = "^2.4.1"
matplotlib = "^3.10.8"
seaborn = "^0.13.2"
scipy = "^1.17.0"
openpyxl = "^3.1.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
deptry = "^0.16.2"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.2.7"
mkdocstrings = {extras = ["python"], version = "^0.26.1"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["patchguru"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"



[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true



[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
